requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
Pillow>=9.0
pandas>=2.0
scikit-image>=0.20
scikit-learn>=1.2
matplotlib>=3.7
scipy>=1.10
albumentations>=1.3
